[core] remove some obsolete functionality and cleanup tests
