+++ src/pymortests/cache.py
@@ -8,7 +8,7 @@ from uuid import uuid4
 from datetime import datetime, timedelta
 
 from pymor.core import cache
-from pymortests.base import TestInterface, runmodule
+from pymortests.base import runmodule
 
 
 SLEEP_DELTA = timedelta(milliseconds=200)
@@ -56,36 +56,34 @@ class IWillBeCopied(cache.CacheableObject):
         return id(self)
 
 
-class TestCache(TestInterface):
-
-    def test_runtime(self):
-        for Class in [IamMemoryCached, IamDiskCached]:
-            r = Class()
-            val = 'koko'
-            int0 = datetime.now()
-            r.me_takey_long_time(val)
-            int1 = datetime.now()
-            r.me_takey_long_time(val)
-            int2 = datetime.now()
-            assert int0 < int1 <= int2
-            delta1 = int1 - int0
-            delta2 = int2 - int1
-            assert delta1>= SLEEP_DELTA, r
-            assert delta2 < delta1, r
-            assert delta2 < 0.5 * SLEEP_DELTA, r
+def test_runtime():
+    for Class in [IamMemoryCached, IamDiskCached]:
+        r = Class()
+        val = 'koko'
+        int0 = datetime.now()
+        r.me_takey_long_time(val)
+        int1 = datetime.now()
+        r.me_takey_long_time(val)
+        int2 = datetime.now()
+        assert int0 < int1 <= int2
+        delta1 = int1 - int0
+        delta2 = int2 - int1
+        assert delta1>= SLEEP_DELTA, r
+        assert delta2 < delta1, r
+        assert delta2 < 0.5 * SLEEP_DELTA, r
 
 
-    def test_region_api(self):
-        with tempfile.TemporaryDirectory() as tmpdir:
-            backends = [cache.MemoryRegion(100), cache.DiskRegion(path=os.path.join(tmpdir, str(uuid4())),
-                                                                    max_size=1024 ** 2, persistent=False)]
-            for backend in backends:
-                assert backend.get('mykey') == (False, None)
-                backend.set('mykey', 1)
-                assert backend.get('mykey') == (True, 1)
-                # second set is ignored
-                backend.set('mykey', 2)
-                assert backend.get('mykey') == (True, 1)
+def test_region_api():
+    with tempfile.TemporaryDirectory() as tmpdir:
+        backends = [cache.MemoryRegion(100), cache.DiskRegion(path=os.path.join(tmpdir, str(uuid4())),
+                                                                max_size=1024 ** 2, persistent=False)]
+        for backend in backends:
+            assert backend.get('mykey') == (False, None)
+            backend.set('mykey', 1)
+            assert backend.get('mykey') == (True, 1)
+            # second set is ignored
+            backend.set('mykey', 2)
+            assert backend.get('mykey') == (True, 1)
 
 
 if __name__ == "__main__":
