+++ src/pymortests/operators.py
@@ -16,13 +16,12 @@ from pymor.operators.numpy import NumpyMatrixOperator
 from pymor.parameters.base import ParameterType
 from pymor.parameters.functionals import GenericParameterFunctional, ExpressionParameterFunctional
 from pymor.vectorarrays.block import BlockVectorSpace
-from pymor.vectorarrays.numpy import NumpyVectorArray, NumpyVectorSpace
-from pymortests.algorithms.stuff import MonomOperator
+from pymor.vectorarrays.numpy import NumpyVectorSpace
 from pymortests.fixtures.operator import (operator, operator_with_arrays, operator_with_arrays_and_products,
-                                          picklable_operator)
+                                          picklable_operator, MonomOperator)
 from pymortests.pickling import assert_picklable, assert_picklable_without_dumps_function
-from pymortests.vectorarray import valid_inds, valid_inds_of_same_length, invalid_inds
-from pymor.core.config import is_windows_platform
+from pymortests.vectorarray import valid_inds, valid_inds_of_same_length
+
 
 def test_selection_op():
     p1 = MonomOperator(1)
